@@ -3,17 +3,17 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@testing-library/jest-dom": "^5.14.1",
+		"@testing-library/jest-dom": "^5.16.4",
 		"@testing-library/react": "^11.2.7",
 		"@testing-library/user-event": "^12.8.3",
 		"@vitejs/plugin-react-refresh": "^1.3.6",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
 		"react-transition-group": "^4.4.2",
-		"styled-components": "^5.3.3",
-		"typescript": "^4.3.5",
-		"vite": "^2.9.12",
-		"vite-plugin-svgr": "^2.1.0",
+		"styled-components": "^5.3.5",
+		"typescript": "^4.7.4",
+		"vite": "^2.9.14",
+		"vite-plugin-svgr": "^2.2.0",
 		"web-vitals": "^1.1.2"
 	},
 	"scripts": {
@@ -40,18 +40,18 @@
 		]
 	},
 	"devDependencies": {
-		"@types/jest": "^26.0.24",
-		"@types/node": "^12.20.19",
-		"@types/react": "^17.0.19",
-		"@types/react-dom": "^17.0.9",
-		"@types/react-transition-group": "^4.4.2",
-		"@types/styled-components": "^5.1.15",
-		"@typescript-eslint/eslint-plugin": "^4.29.2",
-		"@typescript-eslint/parser": "^4.29.2",
-		"eslint": "^7.32.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-prettier": "^4.0.0",
-		"eslint-plugin-react": "^7.24.0",
-		"prettier": "^2.4.1"
+		"@types/jest": "^28.1.4",
+		"@types/node": "^18.0.3",
+		"@types/react": "^18.0.15",
+		"@types/react-dom": "^18.0.6",
+		"@types/react-transition-group": "^4.4.5",
+		"@types/styled-components": "^5.1.25",
+		"@typescript-eslint/eslint-plugin": "^5.30.5",
+		"@typescript-eslint/parser": "^5.30.5",
+		"eslint": "^8.19.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-prettier": "^4.2.1",
+		"eslint-plugin-react": "^7.30.1",
+		"prettier": "^2.7.1"
 	}
 }
