@@ -1,5 +1,5 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import App from 'components/App/App'
 import reportWebVitals from './reportWebVitals'
 import { createGlobalStyle } from 'styled-components'
@@ -29,13 +29,11 @@ const GlobalStyle = createGlobalStyle`
 		margin: 0;
 	}
 `
-
-ReactDOM.render(
+createRoot(document.getElementById('root')!).render(
 	<React.StrictMode>
 		<GlobalStyle />
 		<App />
-	</React.StrictMode>,
-	document.getElementById('root')
+	</React.StrictMode>
 )
 
 // If you want to start measuring performance in your app, pass a function
